Fixed issue #40: mixed type arithmetics lead to a crash while making a db with OMA Jul.2024

+++ omamer/database.py
@@ -1550,7 +1550,7 @@ class DatabaseFromOMABrowser(DatabaseFromOMA):
                     # sequence
                     oma_seq_off = rr["SeqBufferOffset"]
                     seq_len = rr["SeqBufferLength"] - 1
-                    seq = oma_seq_buffer[oma_seq_off : oma_seq_off + seq_len + 1]
+                    seq = oma_seq_buffer[oma_seq_off : np.uint64(oma_seq_off + seq_len + 1)]
                     seq_buffs.append(seq)
 
                     # store protein row
